+++ liubai-frontends/liubai-weixin/miniprogram/packageB/locales/messages/zh-Hans.ts
@@ -80,9 +80,9 @@ export default {
     "great_4": "你正在不断超越自己🥞",
     "great_5": "你是做自己的英雄🍟",
     "create_for_title": "建在哪里？",
-    "create_for_current_chat": "当前会话",
+    "create_for_current_chat": "当前聊天室",
     "create_for_current_group": "当前群聊",
-    "create_for_others": "其他会话或群聊",
+    "create_for_others": "其他",
     "title_for_guests": "只有创建者才能修改标题",
     "updated": "已更新",
     "updated_tip": "请转发↗️，否则他们会不知道",
