+++ liubai-frontends/liubai-weixin/miniprogram/packageB/pages/landing-premium/landing-premium.wxml
@@ -76,4 +76,5 @@
 <custom-navibar 
   height1="{{height1}}" height2="{{height2}}" 
   visible="{{visible}}"
+  always-arrow-back
 ></custom-navibar>
